YARN-2312. Deprecated old ContainerId#getId API and updated MapReduce to use ContainerId#getContainerId instead. Contributed by Tsuyoshi OZAWA (cherry picked from commit f93d2ea27e7554272845fb11d97b3564704834b3)

+++ hadoop-mapreduce-project/hadoop-mapreduce-client/hadoop-mapreduce-client-app/src/main/java/org/apache/hadoop/mapred/WrappedJvmID.java
@@ -23,8 +23,8 @@ package org.apache.hadoop.mapred;
  */
 public class WrappedJvmID extends JVMId {
 
-  public WrappedJvmID(JobID jobID, boolean mapTask, int nextInt) {
-    super(jobID, mapTask, nextInt);
+  public WrappedJvmID(JobID jobID, boolean mapTask, long nextLong) {
+    super(jobID, mapTask, nextLong);
   }
 
 }
